Small state border changes and some resource rebalancing
Shifted Bushehr to Khuzand because it allows Busher to show up on the map as port and replaces the unimportant Bandar-e Mahshahr, it allows for the consolidation Fars of oil and sulfur resources in  Khuzand reducing pop fragmentation and Bushehr was/is mostly Arabic, so it makes culturally sense as well.

Removed the French Basque and Catalonia states.

Redistributed some oil resources and rebalanced fish resources.